tests/wpt/tests/html/syntax/parsing/resources/namespace-sensitivity.dat

#data
<body><table><tr><td><svg><td><foreignObject><span></td>Foo
#errors
(1,6): expected-doctype-but-got-start-tag
(1,56): unexpected-end-tag
(1,60): foster-parenting-character
(1,60): foster-parenting-character
(1,60): foster-parenting-character
(1,60): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     "Foo"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg foreignObject>
|                   <span>
